Implement Solution 1: Create shared constants to eliminate color mapping duplication - Created shared constants file for both Sanity schemas and frontend - Added utility function for color mapping - Updated CTACardBlock and FeatureModules schemas to use shared constants - Eliminates duplicate color mappings across multiple files - Single source of truth for color options and other reusable constants
