+++ studio/src/lib/shared-constants.ts
@@ -1,0 +1,30 @@
+export const ICON_CONTAINER_COLOR_OPTIONS = [
+  { title: 'Red', value: 'red-200' },
+  { title: 'Blue', value: 'blue-200' },
+  { title: 'Bright Yellow', value: 'brightYellow-200' },
+  { title: 'Orange', value: 'orange-200' },
+  { title: 'Lavender', value: 'lavender-200' },
+  { title: 'Wax Flower', value: 'waxFlower-200' },
+  { title: 'Halo Green', value: 'haloGreen-200' },
+]
+
+export const BUTTON_VARIANT_OPTIONS = [
+  { title: 'Default', value: 'default' },
+  { title: 'Secondary', value: 'secondary' },
+  { title: 'Destructive', value: 'destructive' },
+  { title: 'Outline', value: 'outline' },
+  { title: 'Ghost', value: 'ghost' },
+  { title: 'White Ghost', value: 'whiteGhost' },
+  { title: 'White Outline', value: 'whiteOutline' },
+]
+
+export const BACKGROUND_THEME_OPTIONS = [
+  { title: 'Dark', value: 'dark' },
+  { title: 'Creme', value: 'creme' },
+  { title: 'White', value: 'white' },
+]
+
+export const LINK_TARGET_OPTIONS = [
+  { title: 'New Tab', value: '_blank' },
+  { title: 'Same Tab', value: '_self' },
+]
